ai_ref_kits/multimodal_ai_visual_generator/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

openvino-tokenizers==2024.6.0
optimum-intel==1.21.0
numpy==1.26.4
einops==0.8.0  # For Qwen
transformers_stream_generator==0.0.5  # For Qwen
diffusers==0.31.0 # For image generation pipelines
timm==1.0.11  # For exporting InternVL2
torchvision==0.20.1  # For visual language models
transformers==4.46.3 # For Whisper
nncf==2.14.1
accelerate==1.3.0
huggingface_hub==0.28.0
opencv-python==4.11.0.86
pillow==11.1.0
#openvino GenAI (which also installs openvino packages)
openvino_genai==2024.6.0
# for VAD
sherpa-onnx==1.10.41
sounddevice==0.5.1
# for the UI
PySide6==6.8.1